dns: Drop the admin node after initial deployment (SOC-10636)
As it stands, the DNS barclamp is applied with only the admin node in
the dns-server role. As the first part of the proposal step, it adds the
first two nodes (for a total of 3 nodes) so we get multi-master DNS.
Filter out the lead node, which should drop the admin node from the
role.