Unified cloud gating trigger (SOC-10029)
For cloud 8 and 9 the gating jobs should be triggered whenever there is
a new build for Crowbar or Ardana, however to ensure that both medias
have the latest changes the jobs should be executed only when both
medias have the same build number.

This change adds a new job template for triggering the cloud 8 and 9
gating jobs whenever there is a new build for Crowbar and Ardana, it
also checks if they have the same build number and if not mark the build
as unstable and does not run the cloud gating jobs.