openstack-crowbar/ardana: remove dummy lockable resource
Starting with version 2.6 [1], the Jenkins lockable resource plugin
changed its behavior to fail when a non-existing resource label
name is used, even if the resource count is set to zero. This change
invalidates the hack that we use in our declarative pipelines to
conditionally lock a resource based on job input parameters.

Using the workaround suggested here takes care of this problem:
https://github.com/jenkinsci/lockable-resources-plugin/issues/134#issuecomment-517230676

[1] https://github.com/jenkinsci/lockable-resources-plugin/releases/tag/lockable-resources-2.6