workers: Add cron job for cleaning old workspaces
Due to [1] the jenkins jobs are leaving the job's workspace on the worker
node which is causing issues as at some point the worker disk gets full.

This change adds a cron job to the worker that runs daily and deletes the
job workspace directory if it was not accessed on the week.