Skip python-networking-cisco from trackupstream
The repo has been moved to opendev.org/x namespace and the current
tooling no longer able to effective create an updated tarball from it.

If there are customers out there still using this package, we'll need
to update the tooling to be able to create tarballs from
opendev.org/x repos.