Use appropriate Python version for ansible-venv (SOC-10914)
Going forward we want to use Python 3 rather than Python 2, so we
need to ensure that where possible the ansible-venv is created based
upon the desired Python version, where appropriate, rather than the
system default Python version. Note that for Ansible 2.7 this means
we need at least Python 3.5, or Python 2.6.

Additionally the existing run_python_script() prefers Python 3 over
Python 2, but in the case were a system's Python 3 is not viable for
creating the ansible-venv, would still select python3 to run scripts
such as the build_test_package.py script use to build Gerrit based
packages. However if the ansible-venv was created as a Python 2
based virtualenv, and the host system's Python 3 environment doesn't
have a compatible set of Python modules installed to support running
the script under Python 3, you can get errors, even though all of
the required modules have been installed in the ansible-venv.

By ensuring that we use the same python version when creating the
ansible-venv and running scripts we should avoid the above problem,
so long as the ansible-venv has been activated in the runtime
environment.

The combination of the above fixes should improve compatibility of
the manual testing tools with various Linux distributions, including
older versions of openSUSE Leap.