Determine pre-upgrade image automatically
The oncontroller_testpreupgrade() function had hardwired
glance image names for Cloud 8 and Cloud 9 that might not
work, depending on which image URL the Tempest barclamp uses
to create its test images. In that case the function would
be unable to create the Heat stack for upgrade testing and
fail the upgrade on the Cloud 8 side. This commit replaces the
hardwired image name by a heuristic that picks the correct
image name based on the images that are actually present.