openstack-crowbar: enable SSL in test automation for Octavia (SOC-10906)
Enables SSL testing for Octavia in ECP Crowbar deployments.

This commit also changes the Octavia certificate setup logic to
solve the problem described below.

The octavia system group and user are created when the
openstack-octavia package is installed on the controller nodes.
This happens automatically when the octavia barclamp is
applied.

However, setting up the paths for the certificates required for
the two-way SSL amphorae communication needs to be done in advance,
before applying the octavia barclamp, and those paths need to have
their ownership set to the octavia system user/group, which aren't
yet available at that time.

Adding the octavia system user and group is not a good solution,
because it interferes with the barclamp functionality so this change
proposes a different solution to this problem, where the certificate
paths are set up on the controller nodes with root as owner, while
the Octavia barclamp is responsible for switching ownership over
to the octavia user/group during application.