Use external nfsserver for glance if possible
When nfs server was installed on a lonely node, it is registered in
dns as 'nfsserver'. It was used for cinder and rabbit but not for
glance (which used nfs server hosted on the admin node).
For upgrade scenarios, external nfs server is better as it is not
rebooted during the upgrade process.