octavia: adjust tempest filters (SOC-10965)
Adjust the tempest run filters to account for the fact that
when Octavia is enabled, the neutron lbaasv2 API is actually
proxied to the Octavia API and the neutron-lbaas tempest test
cases aren't properly equipped to handle that.

Creates a tempest filter to be run against Octavia with the
lbaasv2-proxy service plugin. The following test cases are
black-listed in that tempest filter, but also black-listed
in qa_crowbarsetup.sh, because they are not compatible with
the Octavia lbaasv2-proxy service plugin:

 - neutron_lbaas.tests.tempest.v2.api.test_health_monitors_non_admin.
 TestHealthMonitors.test_list_health_monitors_two
   - lbaas-proxy doesn't handle the `http_method` attribute correctly
   - see https://jira.suse.com/browse/SOC-10965

 - neutron_lbaas.tests.tempest.v2.api.test_load_balancers_non_admin.
 LoadBalancersTestJSON.test_create_load_balancer_invalid_flavor_field
   - the concept of neutron service flavors doesn't yet have an Octavia
   counterpart in Rocky, so the lbaasv2-proxy service plugin cannot
   transparently convert these attributes
   - see https://blueprints.launchpad.net/octavia/+spec/octavia-lbaas-flavors

 - neutron_lbaas.tests.tempest.v2.api.test_load_balancers_non_admin.
 LoadBalancersTestJSON.test_update_load_balancer_missing*
   - these test cases update an existing load-balancer with an empty
   body (i.e. without actually changing anything)
   - this use case isn't correctly handled by the lbaasv2-proxy
   service plugin