cloud-gating: re-add downstream job URLs
The Blue Ocean UI is still acting up, intermittently failing
to render the "Triggered Builds" section that can be used
to quickly navigate to downstream jobs. To account for this
bug, this commit re-enables printing messages during the Jenkins
pipeline stage, with URLs pointing to triggered downstream jobs.

This commit also changes the way these messages are displayed,
so that they print the URL as a tab label instead of the regular
"Print Message" label.

NOTE: these messages will only show up in the Blue Ocean UI after
a triggered job has completed execution, which means that they
cannot be used to access ongoing downstream jobs.