cloud-mkcloud-gating: disable all mkcloud gating jobs (SOC-10029)
Removes all mkcloud gating jobs. The jobs will be disabled in
Jenkins and kept around until the build log is purged by the
build retention policy, following which the jobs will be
removed.

This is the final phase in the migration to a unified gating job
mechanism. The unified gating jobs that are already in effect
will completely replace the mkcloud legacy gating jobs.