openstack-ardana: ansible role for amphora image setup
We were previously setting up the amphora image on Ardana
deployments only during tempest testing and only if either
the lbaas or heat filters were enabled.

This change moves that logic to its own role and calls it
during regular cloud deployment as well as after a cloud
update, in case the amphora image was updated.