Run ardana-cloud-configure as part of the update workflow (SOC-7028)
Recently a change in tempest-ansible
https://gerrit.prv.suse.net/c/ardana/tempest-ansible/+/6500
 enabled additional tests that require additional resources
to be setup for these tests to execute sucessfully.

However, in the update workflow, the tests were failing since the
tempest-cloud-configure.yml playbook that sets up these resources
were not run.

This change calls the top level ardana-cloud-configure.yml
playbook which calls the tempest-cloud-configure.yml playbook which
creates resources needed for the newly enabled tests.