cloud-generator: fix the swift-object deployment
The swift-object role is by default assigned by Crowbar to all
non-controller nodes. To align unified CI jobs to that, we need to
do the same in the generated Crowbar scenarios. This also takes
some of the load off of controller nodes and reduces the overall
time it takes to re-apply the proposals during update.