qa_crowbarsetup.sh: prepare HA controller node before reboot (SOC-10960)
The `onadmin_rebootcloud` logic from qa_crowbarsetup.sh is not
handling rebooting a cloud node correctly when it is part of
an HA setup. According to the official SLE-HA docs [1], the
controller node needs to be put into standby mode, which
effectively moves all crm resources away from the node and
the pacemaker service has to be stopped before the node can
be safely rebooted.

Not following the correct reboot procedure leads to the node
being fenced half-way through the reboot process, which in turn
causes other problems, such as corrupted root partitions (because
we also have disabled ext journaling on the ECP CI images, to
improve IO performance).

[1] https://documentation.suse.com/sle-ha/15-SP1/single-html/SLE-HA-guide/#pro-ha-maint-shutdown-node