Automate Deployment of Caasp4 using terraform
This playbook provides an automated deployment mechanism
to deploy caasp v4 using terraform on SOC openstack deployment.

This playbook currently runs against the ardana flavored version of
SOC.

You can provide various optional parameters to the playbook:
Some examples:
num_workers to define the number of workers in the cluster.
	- Default value is 1.
num_masters to define the number of masters in the cluster.
	- Default value is 1.

It uses octavia provider for SOC 9 and haproxy provider for SOC 8
deployments.