Caapv4: make k8s use octavia endpoint directly
Creating a Service of type LoadBalancer fails with this message:
"http_method is not a valid option for health monitors of type TCP"

The neuutron lbaas proxy is adding this option which octavia rejects.
Avoid this by having k8s use the octavia endpoint directly instead of
via the neutron lbaas proxy.