devstack: fix the permission of all directories
In 352d58a7afd9e2261e639af78e4fb4c99d8f9f81 of openstack/devstack.git
fixing of permissions on existing directories was dropped. The script
that is changed here relied on that. Adapt it to also take over fixing
the permission of the rest of the directories. The old target of the
chown was /opt/stack/devstack, use the parent instead. This will fix the
following error:

+ ./stack.sh:main:374                      :   mkdir -p /opt/stack/logs
mkdir: cannot create directory ‘/opt/stack/logs’: Permission denied