Ensure cloud_release_number has sane value (SOC-11252)
When SOC/CLM MUs are being tested the ardana-update.yml playbook can
be invoked with an empty value for cloudsource, which breaks recent
changes landed in PR#3726 which use the derived cloud_release value.

This patch ensures that even if cloudsource is empty in such cases
that a valid cloud_release_number is identified from the runtime
env of the job.