Handle ansible_distribution_version not being defined (SOC-11252)
The recent fix for PR#3879 to ensure we select and appropriate cloud
release number when cloudsource is not defined, combined with PR#3880
to handle the renaming of ILO network and BMC network being fixed has
uncovered a follow issue with regard to cloud_release_number handling.

In the case where gathering of facts for an ansible node is disabled,
as is the case with the start-deployer-vm.yml top level playbook, the
use of ansible_distribution_version as a fallback within the PR#3879
fix leads to an error.

This patch addresses this issue by using a default of 'UNKNOWN' for
ansible_distribution_version if it isn't defined so that it will be
easier for any future errors related to this issue being diagnosed.