Ensure ardana-update.yml called with valid cloudsource (SOC-11252)
Only pass update_to_cloudsource value as an override for cloudsource
if a value has been specified for update_to_cloudsource when invoking
the ardana-update.yml playbook both in Jenkins and manual runs. This
allows for the case when running MU testsing where no value is given
for update_to_cloudsource.

Align the logic used to control when the Ardana or Crowbar update and
upgrade playbooks are called in the manual/lib.sh with that used by
the Jenkins pipelines. Namely that update/upgrade should only happen
if update_after_deploy is set.