Tweak integrated deployer/controller VM for physical deploys (SOC-9780)
When doing a baremetal deployment for Ardana with an integrated deployer
the first controller node will be a VM running on the gateway node based
on an appropriate SLES image that has the apparmor pattern installed.

However having apparmor installed on a controller node is not compatible
withe the package set expected for an Ardana controller node, leading to
issues later on after upgrading and rebooting this integrated controller
node from Cloud8/SP3 to Cloud9/SP4.

This patch disables the apparmor service then uninstalls the associated
patterns and unnecessary packages. Note that we leave libapparmor1 as it
is a dependency of other required packages such as rsync.