Gerrit patches for deploy, update and upgrade (SOC-9780)
Add support for being able to build Gerrit test patches for the deploy,
update and upgrade phases, rather than just the initial deploy phase.

To support this we add the following additional job paramaters:
  * update_gerrit_change_ids:
    - A list of Gerrit change ids for the same code stream as the
      specified update_to_cloudsource.
  * upgrade_gerrit_change_ids:
    - A list of Gerrit change ids for the same code stream as the
      specified upgrade_cloudsource.

Similarly there are two new job parameters that can be used to
specify extra repos to be added to the deployer during the update
or upgrade phases:
  * update_extra_repos:
    - A list of extra repos for the same code stream as the
      specified update_to_cloudsource.
  * upgrade_extra_repos:
    - A list of extra repos for the same code stream as the
      specified upgrade_cloudsource.

These repos will be leveraged to inject the results of the relevent
update or upgrade gerrit package builds into the deployer served
repos at the appropriate stage.

Update the Jenkins job, template and Jenkinsfile definitions to
define the new job parameters and leverage them at the appropriate
stages in the Jenings jobs.

Update the manual mode lib.sh with equivalent support for using the
new parameters, which have also been added to the input.yml.example.

Fix the lib.sh get_from_input() shell function to correctly anchor
the requested entry at the start of the line to avoid issues with
'gerrit_change_ids' matching against the update_... and upgrade_...
variants of this entry.