Handle ardana_qe_tests failures better (SOC-9780)
Recently when running iverify tests I have seen issues with the iverify
run completes, likely with failures, and then the subsequent subunit run
to extract the results fails resulting in no ardana_qe_test_results fact
being defined which leads to an ansible error later when checking on the
status of the test that were run.

This patch defines a "default" ardana_qe_test_results fact, with failed
set to 9999999 if we fail to initialise an ardana_qe_test_results due
to the above error scenario.