No 'Cloud' in HOS repo name (SOC-11184)
When trying to deploy a hosdevelcloud8 I see a failure in the step
looking for the repo signing key because it is searching for a key
in directories with 'Cloud' in the name.

However the HOS repo name doesn't have 'Cloud' in it, and hasn't
for a long time if memory serves, since it was renamed during the
final stages of the HOS 8 release.

Instead we should search for "OpenStack" which is common to both
the SOC and HOS repo directory names.