Fix openstack-devstack CI jobs (SOC-11405)
Update the job definitions for the openstack-devstack-ipv{4,6} jobs
to use the Leap 15.2 image when creating a cleanvm instance, and to
manage the cleanvm vm on the ci1-opensuse system rather than locally.

Update the qa_devstack.sh script to specify memory_tracker rather
than peakmem_tracker when defining the devstack local.conf config
for the master branch.

Additionally the upstream devstack scripts now expect an env var,
PYTHON, defined specifing the appropriate python command to use
when running scripts such as outfilter.py, which are not marked
as being executable.