Fix amphora image rpm selection process (SOC-11407)
The linux find command returns the list of files found in a directory
in the order they are found in the directory listing which is usually
some form of inode order, and not necessarily lexically sorted.

This patch fixes the issue by including a 'sort --version-sort' in
the shell command pipeline before the 'tail -n1' so that the list of
RPM packages is version sorted.