Switch Cloud8 to use legacy ISO server (SOC-11466)
Recent changes to the mirroring policies for ibs-mirror.prv.suse.net
mean that it no longer mirrors the SLE 12 SP3 ISO that is needed by
the SOC 8 CLM cobbler (PXE Boot) re-imaging process.

I have manually setup a mirror of the necessary ISO on the legacy
CI artifacts server for Ardana, ardana.ci.prv.suse.net, with the
same directory hierarchy, so we can just switch the hostname in
the URL to fix this.

Also updated an occurrence of the old suse.provo.cloud version of
the name for the legacy CI artifacts server.