Move from travis to Jenkins
Travis shut down or something, use this chance to move the tests to
the existing Jenkins.

This adds a Jenkinsfile to call each of the targets in make test
individually so that the output is easier to read in Jenkins.

Replace pip packages with zypper packages in make install.

Add perl-JSON to make install. It was was missed until now.