Allow changing the HA sync mark timeout multiplier
This was already supported in the jobs running with Openstack Heat
stacks. Implement the same thing when running with mkcloud.

Both implementations are now changed to by default set it to 2.0, before
the default was to not change it.

I leave the implementation used by Heat stacks as is to reduce risk, as
a result there it will be set it twice.